user-service/requirements.txt
uvicorn
fastapi[standard]
python-dotenv
beanie
pydantic[email]
datetime
motor
